moved all boundary related commands into a new js file
