+++ frontend/src/js/commands/boundary.commands.js
@@ -1,7 +1,9 @@
 import { LaunchTask } from "../util/firefly";
 import { ParseBackendRegion } from "../util/region";
 import * as HardwareRegions from "../util/hardwareregion";
-import { setBoundaryRegions } from "../actions/viewer.actions";
+import { clearLayer, drawDS9Regions, setBoundaryRegions } from "../actions/viewer.actions";
+import { validateParameters } from "../util/command";
+import { addErrorToHistory } from "../actions/terminal.actions";
 
 import store from "../store";
 
@@ -16,6 +18,54 @@ const parseCCD = data => {
 
         return new HardwareRegions.CCD(name, regions);
     });
+}
+
+// Displays the boundary regions on a viewer
+export const showBoundary = params => {
+    const valid = validateParameters(params, store.getState());
+    if (valid !== null) {
+        store.dispatch(addErrorToHistory("Bad parameters: " + valid));
+        return;
+    }
+    const viewerID = params.viewer_id;
+    const viewers = store.getState().viewers;
+    const viewer = viewers[viewerID];
+    const regionLayer = "BOUNDARY";
+
+    if (!viewer.boundaryRegions){
+        const err = "Boundary not fetched for this image."
+        console.log(err);
+        return;
+    }
+
+    store.dispatch(clearLayer(viewerID, regionLayer));
+    const regions = viewer.boundaryRegions;
+    const opts = {
+        color: "red",
+        width: 1
+    };
+
+    // Draw the boundary regions
+    store.dispatch(drawDS9Regions(viewerID, regionLayer, regions, opts));
+}
+
+export const hideBoundary = params => {
+    const valid = validateParameters(params, store.getState());
+    if (valid !== null) {
+        store.dispatch(addErrorToHistory("Bad parameters: " + valid));
+        return;
+    }
+    const viewerID = params.viewer_id;
+    const viewers = store.getState().viewers;
+    const viewer = viewers[viewerID];
+    const regionLayer = "BOUNDARY";
+    // TODO: move boundary functions to a separate js file
+    if (!viewer.boundaryRegions){
+        const err = "Boundary not fetched for this image."
+        console.log(err);
+        return;
+    }
+    store.dispatch(clearLayer(viewerID, regionLayer));
 }
 
 // Parse CCD with overscan format
