+++ frontend/src/js/commands/viewer.commands.js
@@ -6,6 +6,8 @@ import { addErrorToHistory } from "../actions/terminal.actions";
 
 import store from "../store";
 
+// TODO: ADD validateParameters!!!
+
 // Clears all region layers on a viewer
 export const clearViewer = params => {
     store.dispatch(ViewerActions.clearViewer(params.viewer_id));
@@ -32,52 +34,4 @@ export const deleteViewer = params => {
 // Loads a new image into a viewer
 export const loadImage = params => {
     store.dispatch(ViewerActions.loadImage(params.viewer_id, params.url));
-}
-
-// Displays the boundary regions on a viewer
-export const showBoundary = params => {
-    const valid = validateParameters(params, store.getState());
-    if (valid !== null) {
-        store.dispatch(addErrorToHistory("Bad parameters: " + valid));
-        return;
-    }
-    const viewerID = params.viewer_id;
-    const viewers = store.getState().viewers;
-    const viewer = viewers[viewerID];
-    const regionLayer = "BOUNDARY";
-
-    if (!viewer.boundaryRegions){
-        const err = "Boundary not fetched for this image."
-        console.log(err);
-        return;
-    }
-
-    store.dispatch(ViewerActions.clearLayer(viewerID, regionLayer));
-    const regions = viewer.boundaryRegions;
-    const opts = {
-        color: "red",
-        width: 1
-    };
-
-    // Draw the boundary regions
-    store.dispatch(ViewerActions.drawDS9Regions(viewerID, regionLayer, regions, opts));
-}
-
-export const hideBoundary = params => {
-    const valid = validateParameters(params, store.getState());
-    if (valid !== null) {
-        store.dispatch(addErrorToHistory("Bad parameters: " + valid));
-        return;
-    }
-    const viewerID = params.viewer_id;
-    const viewers = store.getState().viewers;
-    const viewer = viewers[viewerID];
-    const regionLayer = "BOUNDARY";
-    // TODO: move boundary functions to a separate js file
-    if (!viewer.boundaryRegions){
-        const err = "Boundary not fetched for this image."
-        console.log(err);
-        return;
-    }
-    store.dispatch(ViewerActions.clearLayer(viewerID, regionLayer));
 }
